+++ Assets/1 Objects/Items/Item.cs
@@ -46,7 +46,7 @@ public class Item : Block
             return;
         }
 
-        int cellSize = GridManager.Instance.CellSize;
+        float cellSize = GridManager.Instance.CellSize;
         
         // Check if any targeted Cell is already Blocked
         for (int i = 0; i < blockLayoutOffset.Count; ++i)
@@ -70,7 +70,7 @@ public class Item : Block
 
     private void UpdateBlockCells(Vector3 newPos)
     {
-        int cellSize = GridManager.Instance.CellSize;
+        float cellSize = GridManager.Instance.CellSize;
         Vector3 offset;
 
         // Empty All old positions
